Add generation flags
