+++ src/commands/generate/__tests__/simpleAstTest.ts
@@ -22,9 +22,9 @@ import { directory } from "tempy";
 import { createHashableTypeName } from "../../../utils/hashingUtils";
 import { DEFAULT_TYPE_GENERATION_FLAGS } from "../../../__tests__/utils/constants";
 import { generateError } from "../errorGenerator";
-import { generateNonThrowingService } from "../nonThrowingServiceGenerator";
+import { generateNonThrowingService } from "../generateNonThrowingService";
 import { SimpleAst } from "../simpleAst";
-import { generateThrowingService } from "../throwingServiceGenerator";
+import { generateThrowingService } from "../generateThrowingService";
 import { generateEnum } from "../typeGenerator";
 
 describe("simpleAst", () => {
