+++ src/commands/generate/generateNonThrowingService.ts
