+++ src/commands/generate/generator.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { IConjureDefinition, IServiceDefinition, ITypeDefinition } from "conjure-api";
+import { IConjureDefinition, ITypeDefinition } from "conjure-api";
 import * as fs from "fs-extra";
 import * as _ from "lodash";
 import * as path from "path";
@@ -23,37 +23,18 @@ import { ITypeGenerationFlags } from "../../types/typeGenerationFlags";
 import { directoryNameForType } from "../../utils/fileUtils";
 import { createHashableTypeName, disassembleHashableTypeName } from "../../utils/hashingUtils";
 import { generateError } from "./errorGenerator";
-import { generateNonThrowingService } from "./nonThrowingServiceGenerator";
+import { generateNonThrowingService } from "./generateNonThrowingService";
 import { SimpleAst } from "./simpleAst";
-import { generateThrowingService } from "./throwingServiceGenerator";
+import { generateThrowingService } from "./generateThrowingService";
 import { generateType } from "./typeGenerator";
-
-const assertNoServiceNamesConflicts = (services: IServiceDefinition[]): void => {
-    const packageToServiceNamesMap: Map<string, Set<string>> = new Map();
-    services.forEach(service => {
-        const { name: serviceName, package: packageName } = service.serviceName;
-        if (!packageToServiceNamesMap.has(packageName)) {
-            packageToServiceNamesMap.set(packageName, new Set());
-        }
-
-        packageToServiceNamesMap.get(packageName)!.add(serviceName);
-    });
-
-    Array.from(packageToServiceNamesMap.entries()).forEach(([packageName, serviceNames]) => {
-        serviceNames.forEach(serviceName => {
-            if (serviceNames.has(`${serviceName}WithErrors`)) {
-                throw new Error(
-                    `Found service name conflict in ${packageName}. Conflict: ${serviceName}, ${serviceName}WithErrors`,
-                );
-            }
-        });
-    });
-};
+import { IServiceGenerationFlags } from "../../types/serviceGenerationFlags";
+import { validateServiceNames } from "./validators/validateServiceNames";
 
 export async function generate(
     definition: IConjureDefinition,
     outDir: string,
     typeGenerationFlags: ITypeGenerationFlags,
+    serviceGenerationFlags: IServiceGenerationFlags,
 ) {
     // Create project structure
     const knownTypes = computeKnownTypes(definition.types);
@@ -70,7 +51,8 @@ export async function generate(
         ),
     );
 
-    assertNoServiceNamesConflicts(definition.services);
+    // Validate service names if user wants to generate throwing an non-throwing services
+    validateServiceNames(definition.services, serviceGenerationFlags);
 
     const knownDefinitions = Array.from(knownTypes.keys())
         .map(disassembleHashableTypeName)
@@ -89,15 +71,22 @@ export async function generate(
     const promises: Array<Promise<any>> = [];
     const simpleAst = new SimpleAst(outDir);
 
-    definition.services.forEach(serviceDefinition =>
-        promises.push(generateThrowingService(serviceDefinition, knownTypes, simpleAst, typeGenerationFlags)),
-    );
-    definition.services.forEach(serviceDefinition =>
-        promises.push(generateNonThrowingService(serviceDefinition, knownTypes, simpleAst, typeGenerationFlags)),
-    );
+    if (serviceGenerationFlags.generateThrowingServices) {
+        definition.services.forEach(serviceDefinition =>
+            promises.push(generateThrowingService(serviceDefinition, knownTypes, simpleAst, typeGenerationFlags)),
+        );
+    }
+
+    if (serviceGenerationFlags.generateNonThrowingServices) {
+        definition.services.forEach(serviceDefinition =>
+            promises.push(generateNonThrowingService(serviceDefinition, knownTypes, simpleAst, typeGenerationFlags)),
+        );
+    }
+
     definition.types.forEach(typeDefinition =>
         promises.push(generateType(typeDefinition, knownTypes, simpleAst, typeGenerationFlags)),
     );
+
     definition.errors.forEach(errorDefinition =>
         promises.push(generateError(errorDefinition, knownTypes, simpleAst, typeGenerationFlags)),
     );
